+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Hardcoded `sha512` digest in `pkcs:pbkdf2/4` function.
+
 ## [0.1.0] - 2020-12-17
 ### Added
-- PBKDF2 algorithm
+- PBKDF2 algorithm.